Rearranged mappers, set spellbook id in client.
